refactor: Consolidate pipeline processor implementation and tests
1. Replaced the original pipeline processor with the refactored version
2. Added conditional multithreading checks to the original tests
3. Updated create_config to handle dictionary format for weights
4. Removed redundant refactored test file
5. Removed redundant refactored pipeline file

This commit completes the pipeline refactoring by consolidating the code into a single implementation and test suite, making it easier to maintain going forward.
